requirements.txt
selenium==4.6.0
requests==2.28.1
python-telegram-bot==20.0a0
promise-keeper==1.0.0
lazy-streams==1.0.0
